# Improve performance of lazy validation message formatting