@bluetech 2019-05-23
For future reference, same benchmark was used:  https://gist.github.com/bluetech/cdd3b3d6d9d8cfc7b1dfee3c6482679d