@carltongibson 2019-05-24
Argh! Don't do that. 🙂

I'm quietly getting knocked sideways by life over here, so any help getting the 3.10 issues in is very welcome. 

I've got the schema docs to finish, which I'm on, but beyond that I keep seeing the number of things to pull in, vs the time I'm able to make for it, and I'm starting to panic a little. 

(Was going to say that in private, but TBH, given how OSS works, I think these things are better off said.) 

Quick look at this: it's seems fine, but I haven't looked deeply. Happy to trust @rpkilby's review.