@johnthagen 2019-07-28
Just wanted to say that this fix improved the speed of some deeply nested serializers in our project by 60%! ❤️ **Huge** real world improvement in DRF 3.10.

Thank you all!